+++ docs/components/Footer.vue
@@ -1,16 +1,22 @@
 <script setup lang="ts">
-const config = useRuntimeConfig().public
+const links = inject<Ref<any[]>>('links')
 </script>
 
 <template>
-  <footer class="flex items-center gap-1.5 mt-12">
-    <NuxtLink :to="`https://github.com/Dave136/vue-email/releases/tag/v${config.version}`" target="_blank" class="inline-flex">
-      <UBadge :label="`v${config.version}`" variant="subtle" />
-    </NuxtLink>
-
-    <div class="flex-1 flex items-center justify-end gap-1.5 -my-1 lg:hidden">
-      <USocialButton to="https://discord.gg/Z2RW7qvS3Q" target="_blank" icon="i-simple-icons-discord" />
-      <USocialButton to="https://github.com/Dave136/vue-email" target="_blank" icon="i-simple-icons-github" />
+  <div v-if="$route.path !== '/playground'" class="w-full h-px bg-gray-200 dark:bg-gray-800 flex items-center justify-center">
+    <div class="px-4">
+      <Logo class="w-5 h-5" />
     </div>
-  </footer>
+  </div>
+
+  <UFooter :links="links" :ui="{ bottom: { container: 'lg:py-4' } }">
+    <template #left>
+      <span class="text-sm"> Published under <NuxtLink to="https://github.com/Dave136/vue-email" target="_blank" class="underline"> MIT License </NuxtLink> </span>
+    </template>
+
+    <template #right>
+      <UColorModeButton v-if="!$colorMode.forced" />
+      <UButton aria-label="Vue Email on GitHub" icon="i-simple-icons-github" to="https://github.com/Dave136/vue-email" target="_blank" color="gray" variant="ghost" />
+    </template>
+  </UFooter>
 </template>