+++ playgrounds/nuxt/server/api/test.get.ts
@@ -2,8 +2,21 @@ import { useCompiler } from '#vue-email'
 
 export default defineEventHandler(async () => {
   try {
-    const template = await useCompiler('github-access-token.vue', {
-      username: 'flowko',
+    const template = await useCompiler('Components:translation.vue', {
+      props: {
+        username: 'flowko',
+      },
+      i18n: {
+        defaultLocale: 'en',
+        translations: {
+          en: {
+            message: 'Welcome to dashboard {username}',
+          },
+          es: {
+            message: 'Bienvenido al panel {username}',
+          },
+        },
+      },
     })
 
     if (!template) {