+++ src/nuxt/runtime/templates/vue-email.ts
@@ -1,8 +1,9 @@
-import { VueEmailPlugin } from 'vue-email'
+import { VueEmailPlugin, VueEmailPluginOptions } from 'vue-email'
 import {defineNuxtPlugin, useRuntimeConfig } from "#app"
 
 export default defineNuxtPlugin((nuxtApp) => {
   const config = useRuntimeConfig()
+  const options = config.public['vueEmailOptions'] as VueEmailPluginOptions
 
-  nuxtApp.vueApp.use(VueEmailPlugin, config.public.vueEmailOptions)
+  nuxtApp.vueApp.use(VueEmailPlugin, options)
 })