+++ src/plugin/index.ts
@@ -3,13 +3,12 @@ import type { Plugin } from 'vue'
 import * as components from '../components'
 import type { VueEmailPluginOptions } from '../types'
 import { config } from '../config'
-import { assign } from '../utils/assign-deep'
+import { deepmerge } from '../utils/assign-deep'
 
 export const VueEmailPlugin: Plugin = {
   install(app, options: VueEmailPluginOptions = {}) {
-    if (options) assign(config, options)
+    if (options) deepmerge<VueEmailPluginOptions>(config, options)
 
-    // Register components
     Object.entries(components).forEach(([name, component]) => {
       app.component(name, component)
     })