+++ src/types/compiler.ts
@@ -1,4 +1,44 @@
-import type { VueEmailPluginOptions } from './index'
+export interface VueEmailPluginOptions {
+  /**
+   * The base URL of your website.
+   * @default null
+   * @example
+   * ```ts
+   * baseUrl: 'https://example.com'
+   * ```
+   * @see https://vue-email.net/getting-started/installation#options
+   */
+  baseUrl?: string | null
+  /**
+   * Provide translations for your templates.
+   * @see
+   */
+  i18n?: i18n
+  [key: string]: any
+}
+
+export interface i18n {
+  /**
+   * The default locale if none is provided.
+   * It is also used as fallback if translation does not exist in requested locale.
+   */
+  defaultLocale: string
+  /**
+   * @see https://vue-i18n.intlify.dev/
+   * @example
+   * ```ts
+   * translations: {
+   *   en: {
+   *     hello: "Hello"
+   *   },
+   *   sv: {
+   *     hello: "Hej"
+   *   }
+   * }
+   * ```
+   */
+  translations?: Record<string, Record<string, string>>
+}
 
 export interface DefineConfigFunctions {
   render: (name: string, options?: RenderOptions) => Promise<string>
@@ -23,6 +63,7 @@ export interface Options {
 
 export interface RenderOptions {
   props?: Record<string, unknown>
+  i18n?: i18n
 }
 
 export type DeepRequired<T> = Required<{ [P in keyof T]: DeepRequired<T[P]> }>