+++ docs/content/changelog.md
@@ -1,6 +1,10 @@
 ---
 title: Changelog
 description: Discover the latest Vue Email updates.
+links:
+  - label: Github Releases
+    icon: i-simple-icons-github
+    to: https://github.com/Dave136/vue-email/releases
 ---
 
 :change-log